requirements.txt
Flask==0.12
PyYAML==5.1
google-cloud-dialogflow==2.4.0